Annotation refactor
